% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Testing roxygen2 documentation. This is a title.}
\usage{
hello()
}
\value{
A string.
}
\description{
Add a space to separate title from description body.
Rebuild documentation with devtools::document(). See http://r-pkgs.had.co.nz/man.html
}

