% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rtsamp}
\alias{rtsamp}
\title{Sample RT data}
\format{A data frame with 179 observations and 2 variables:
\describe{
  \item{RT}{Reaction time, in milliseconds.}
  \item{hit}{0 or 1, whether the target was hit or not.}
}}
\usage{
rtsamp
}
\description{
A dataset from Adrian with reaction time data. Also includes NAs.
}
\keyword{datasets}

